Keep keybaord closed after Google billing
For some reason the Google Play Billing dialog is causing the keyboard
to open when it closes. I don't know what's causing it, and this was the
only way I found to keep it closed, but it definitely results in some
jank as the keyboard pops open and then is immediately closed, so it's
far from ideal.