Make search bar centered vertically
Matters when the app bar is a different color because before this change
the search bar was up against the differently-colored app bar, but it
had 16dp of space below it. This ensures it is centered in these cases,
but does cause it to appear a bit uncentered when the app bar is not
colored (like with the light theme).