Only show onboarding activity on first onCreate call.
This is to avoid an issue observed with emulators where onCreate would
get called twice when starting the app. This may just be an emulator
issue since an "identical" physical device did not show the issue, but
adding this check since it seems harmless and helps protect us from this
issue.