Activate user selected Plus theme after upgrade
The removed code was added to avoid a bug where the selected theme/icon
would not reset after the user declined to upgrade. I wasn't able to
recreate that bug in my testing though even with the changes in this
commit. Furthermore, the changes to fix that bug broke the behavior
where the user's selected Plus theme would be activated after they
upgraded. This commit addresses that.