Remove unnecessary direct setting of starred on episode object
We don't need this here because it's done in the MediaSessionManager. We
ideally wouldn't need to do it there either, but it seems that our up
next queue does not listen to changes to the episodes in it (and the
episodes stored in the up next queue handler are used in many places
throughout the app).