Fix Play button color not updating after changing filter
PlayButton.setButtonType is used to update the button appearance.
This method did an early return without checking if the color was
updated.

To fix the issue a check was added to check if the previous color is
same as current color, if no then we update the button. To keep track
of the previous color an additional variable was added. PlayButton
now keeps track of the previous selected color.