Simplify application coroutine scope by just providing it at the repository level
Providing it at the app-level feels a bit purer, but it adds a fair bit
of complexity with DI since we then have to provide it separately for
the phone, automotive, and watch apps. So just going with the simpler
approach for now and if we need to inject it separately for the
different apps later for some reason, we can always add the DI
complexity then.