Migrate addNoMediaFile() method to Kotlin
For simplicity of the migration the method was changed to be non-static.

The null check form the folder parameter got removed. All usages of the addNoMediaFile() method are internal to the FileStorage class and every invocation uses a File created directly through a constructor so there is no possibility for it to be null.