Log MultiSelect events instead adding breadcrumbs
Because `toolbarActions` are evaluated before crashlogging SDK, we have crashing app (in debug) informing us, that logging breadcrumbs at this moment shouldn't happen, as it does nothing (SDK is not yet initialized).

Instead of adding breadcrumbs, I propose logging those events as, in the end, it'll be soon available via encrypted logs.