Build all prototype apps and prepare a combined comment
Because the project is modularized, it makes sense to build all prototype builds in a single Gradle build, rather than running 3 Buildkite jobs and occupy CI for maybe a little improvement.

Also, prototype builds is not the longest CI job when verifying a commit, so building all prototype builds in one job is not extending the overall waiting time on a PR to finish all checks.