Create Gravatar module to be injected where requiered
We don't want the Gravatar SDK to land in the wear/automotive builds. The Gravatar module will be the only one that depends on the Gravatar SDK so that we can inject it only in the "app" build. Similar to what we are doing with "media-ffmpeg" and "media-noop" modules.

In this commit, "wear" and "automotive" builds are broken. We need to add the "gravatar-noop" module and inject it into those modules.