Point CI toolkit to WIP version that bypasses caching when AWS fails
This is to workaround the kind of failure that @geekygecko reported
internally in p1740374258002719/1740354379.359239-slack-CC7L49W13 and
can be seen in CI at
https://buildkite.com/automattic/pocket-casts-android/builds/10679#019535dd-890f-413d-9c9c-2e0494ee8fe5

The failure was due to the logic to download the automation tooling Ruby
gems cache failing silently. The failure itself was due to the AWS CLI
somehow failing to authenticate with AWS.

The WIP branch doesn't fix the issue. It only introduces additional
checks to surface failures to authenticate with AWS and bypass caching.