Disable running `lint_task` during `finalize_release`
The `finalize_release` lane now runs on CI, but uses the `tumblr-metal` agent to do so because we need a trusted agent with permissions to `git push`.

Those agents are not supposed to be used to run `gradle` tasks or doing build stuff, but rather only to do small commit tasks (like version bump or translations file download). So we shouldn't make the `android_download_translations` task run the gradle `lint_task: 'aggregatedLintRelease'` on those CI agents.

This commit temporarily disable the linting task to workaround the issue that the `tumblr-metal` agent can't even download the gradle binary (because it's behind a proxy and gradle is not provided the values of `HTTP_PROXY` env vars as `-Dhttp.proxyHost` system properties, so it fails when trying to download it anyway) and failed the CI build.

In a future PR we will split the lane in 2 to run the translations download on one lane (run on `tumblr-metal` so it can push), then let CI do the linting on a proper `android` agent, then have a second lane do the rest of the release finalization. This will require an update of the release scenario though, so would only be applicable for the next release