Deps: Group kotlin & ksp dependabot updates
With this configuration Dependabot should update both, the 'Kotlin' and
'KSP' related dependencies in one go, on a single Dependabot related PR.

FYI: Using the 'applies-to key' is optional. If the 'applies-to key' is
absent from a set of grouping rules, it defaults to 'version-updates'
for backwards compatibility. I chose to use 'applies-to' in order to be
explicit about this and avoid any confusion related to the defaults.

For more info see: https://docs.github.com/en/code-security/
dependabot/dependabot-version-updates/
configuration-options-for-the-dependabot.yml-file#groups

------------------------------------------------------------------------

PS: With this configuration in place, the 'ksp' update can be now
removed from the ignored list.