Automate publishing of GitHub Release for betas
Now that we create all GitHub Releases as draft, even the betas, we need to have those draft GitHub Releases be published once the Release Manager has smoke-tested the beta and is ready to start the initial rollout.

In order to avoid the publication of the GitHub Release to be a _separate_ step in the ReleasesV2 scenario, especially given there's already a step in the scenario that triggers a CI build for updating the GPC rollouts, we're making the `.buildkite/release-pipelines/update-rollouts.yml` pipeline be able to _also_ call `bundle exec fastlane publish_github_release` as well if `PUBLISH_DRAFT_GITHUB_RELEASE=true` env var was passed.

This means that we don't need the `release-pipelines/publish-release.yml` CI pipeline anymore.

Instead, in the ReleasesV2 scenario, we'll be replacing tasks that were calling this `publish-release.yml` pipeline, which were all called right before one that was triggering `update-rollout.yml` already, to pass `PUBLISH_DRAFT_GITHUB_RELEASE=true` in the call to the task triggering `update-rollout.yml` instead, to kill two birds with one stone.

(Note that some tasks in the scenario will still call `update-rollout.yml` with `PUBLISH_DRAFT_GITHUB_RELEASE=false` when those are about _increasing_ the rollout of an APK that has already published (e.g. going from 10% to 50%, as opposed to going from 0% to 1%)