Build: Add missing ignored variants on dependency analysis
Although 'release' was already defined as an ignored variant,
'debugProd' and the newly added 'prototype' variants aren't yet. This
change makes sure that only the 'debug' variant is active for dependency
analysis purposes, which also makes the 'buildHealth' run faster without
the need to analyze those extra variants.

FYI: And actually, the 'prototype' variant is also causing failures on
CI due the 'google-services.json' file missing. This change fixes that.